+++ app/loadLilyParser.js
@@ -1,5 +1,5 @@
 
-import * as lilyParser from "../inc/lilyParser";
+import jisonWrapperWorker from "comlink-loader!../inc/jisonWrapper";
 
 
 
@@ -17,14 +17,20 @@ export default async function load () {
 
 		const t1 = performance.now();
 
-		// mute jison logs during grammar loading
+		/*// mute jison logs during grammar loading
 		lilyParser.hookJisonPrint();
 
 		parser = lilyParser.createParser(grammar);
 		if (!parser)
 			console.error("lily parser create failed.");
 
-		lilyParser.hookJisonPrint((...args) => console.log("[JISON]", ...args));
+		lilyParser.hookJisonPrint((...args) => console.log("[JISON]", ...args));*/
+		//const lilyParser = lilyParserWorker();
+		//console.log("lilyParser:", lilyParser);
+		const jisonWrapper = new jisonWrapperWorker();
+
+		parser = await new jisonWrapper.Parser(grammar);
+		//console.log("parser:", parser, parser.parse);
 
 		const t2 = performance.now();
 		console.debug("lilyParser loading cost:", t1 - t0, t2 - t1);