+++ inc/lilyParser/utils.ts
@@ -57,6 +57,22 @@ class FractionNumber {
 };
 
 
+const replaceSourceToken = (source: string, token: string): string => {
+	let placeholder = "";
+
+	if (token.length < 4)
+		placeholder = Array(token.length).fill(" ").join("");
+	else
+		placeholder = "%{" + Array(token.length - 4).fill("-").join("") + "%}";
+
+	let result = source;
+	while (result.includes(token))
+		result = result.replace(token, placeholder);
+
+	return result;
+};
+
+
 
 export {
 	WHOLE_DURATION_MAGNITUDE,
@@ -64,4 +80,5 @@ export {
 	lcm,
 	lcmMulti,
 	FractionNumber,
+	replaceSourceToken,
 };